### 2018-06-17 #761 @labarba
@drvinceknight In several places (repo, docs), you have a note like "to do sophisticated equilibria computation you should use gambit" — could you explain how this software, which you say computes equilibria in 2-player games only, would be useful in a research scenario?
