### 2018-06-18 #761 @drvinceknight
> After inspecting the software repository and the documentation, I get the clear impression that this package has more of a pedagogical value than a research application. You recently reviewed one of our first submissions to JOSE (thanks!) ... that software originally came to JOSS—have a look at the pre-review discussion: #212
> Do you think your submission here is in a similar situation? (That would be my impression.)

I actually was hesitating in my choice for this one: I'll happily resubmit this over to JOSE :)
