### 2018-06-18 #761 @labarba
👍  I do think you should include a couple of worked-out examples, in that case. For JOSE, one thing we want is to make it easier for someone else to adopt the software/materials, or for someone to use them for self-learning. Worked examples are important for that.
