### 2018-06-18 #761 @drvinceknight
> I do think you should include a couple of worked-out examples, in that case. For JOSE, one thing we want is to make it easier for someone else to adopt the software/materials, or for someone to use them for self-learning. Worked examples are important for that.

Yup of course, I'm at a conference at the moment but will reshape the paper for JOSE. In terms of worked examples, do you mean in `paper.md`? The library [documentation](http://nashpy.readthedocs.io/en/stable/) has quite an extensive set of examples.

A thought: I've actually used Nashpy for a game theory course which uses software in a lot of other ways as well that I was thinking as writing up as a paper for JOSE also. (Use of gh-pages as content delivery, heavy reliability on jupyter notebooks with a testing framework, bespoke static site generator, as well as sphinx backed documentation of "how to lead the course".) My intent is that the paper on Nashpy and the course would not be the same.
