### 2018-06-25 8917d00
Refactor paper for submission to JOSE.
Add an example use case.
