+++ paper.md
@@ -6,7 +6,7 @@ tags:
   - economics
   - computer science
   - game theory
-  - equilbrium
+  - equilibrium
 authors:
   - name: Vincent Knight
     orcid: 0000-0002-4245-0638
@@ -42,22 +42,45 @@ scientific Python stack (numpy and scipy [@scipy]) thus it is portable. Nashpy
 currently implements 3 algorithms for the computation of equilibria (currently
 only for 2 player games) and is extensively documented, including theoretic
 reference material on the algorithms:
-[nashpy.readthedocs.io](http://nashpy.readthedocs.io/). Furthermore, the
+[nashpy.readthedocs.io](http://nashpy.readthedocs.io/). This documentation
+coupled with the readability of Python make it a particularly effective teaching
+tool as students can inspect the code to reinforce
+their understanding of the algorthms.
+Furthermore, the
 software is automatically tested using a combination of unit, integration and
-property based tests with 100% coverage.
+property based tests with 100% coverage. All the documentation is doctested and
+in fact the example in this paper is as well.
+
+An often used game in beginner game theory classes is referred to as the *battle
+of the sexes* and here is an example of how to use ``Nashpy`` to obtain the
+equilibria:
+
+```
+>>> import nashpy as nash
+>>> import numpy as np
+>>> A = np.array([[2, 0], [0, 1]])
+>>> B = np.array([[1, 0], [0, 2]])
+>>> game = nash.Game(A, B)
+>>> for eq in game.support_enumeration():
+...     print(eq)
+(array([ 1.,  0.]), array([ 1.,  0.]))
+(array([ 0.,  1.]), array([ 0.,  1.]))
+(array([ 0.666...,  0.333...]), array([ 0.333...,  0.666...]))
+
+```
 
 ``Nashpy`` is designed to be used by researchers and students in courses in the
 fields of mathematics, computer science and/or economics. It is already
 currently being used in a final year course at Cardiff University.
-Due to the fact that the code is written entirely in Python and is open source,
-this makes it a positive teaching tool as students can read and understand
-implementation of the algorithms.
 ``Nashpy`` has been archived to Zenodo with the linkd DOI:
 [@zenodo].
 
+
+```
+
 # Acknowledgements
 
-I acknowledge code contributions from Ria Baldevia as well as many helpful
+We acknowledge code contributions from Ria Baldevia as well as many helpful
 discussions with Nikoleta Glynatsi.
 
 # References
