### 2018-06-26 97da11b
Remove referral to game as "Battle of the Sexes"
This was requested here:
https://github.com/openjournals/jose-reviews/issues/18#issuecomment-400227063
