+++ docs/conf.py
@@ -20,7 +20,7 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('../src/'))
-import nash
+import nashpy as nash
 
 
 # -- General configuration ------------------------------------------------
