### 2018-07-15 31dc9c1
Refactor paper for un withdrawal to JOSS.
