### 2018-07-15 #761 @drvinceknight
Dear editors,

I would like to un-"withdraw" this paper, please let me know if it's simpler if I were to resubmit it perhaps?

This follows a discussion at JOSE https://github.com/openjournals/jose-reviews/issues/18 were I believe it has become clear that a paper about this software library is not within the scope of JOSE.

I apologise for the confusion/mistake.

@labarba regarding your queries prior to withdrawal:

> @drvinceknight In several places (repo, docs), you have a note like "to do sophisticated equilibria computation you should use gambit" — could you explain how this software, which you say computes equilibria in 2-player games only, would be useful in a research scenario?

I have modified the paper to hopefully make this clearer, the following is now included:

```
# Statement of need

Access to these algorithms is non trival, an example of this includes the
modelling of healthcare decisions [@knight2017measuring] where a bespoke
theoretic result was used to design a specific algorithm for the computation of
equilibria. Easily accessibly software would make that research more
straightforward as no new algorithm would need to be implemented.

The most mature piece of software available for the computation of equilibria
is **Gambit** [@mckelvey2006gambit]. Gambit includes a python wrapper to its
core C functionality however is not currently portable for example
Windows is not supported. There does exist a web interface with a Gambit back
end: [Game theory
explorer](http://gte.csc.liv.ac.uk/index/index.html#document-documentation)
however this is not practical for reproducible research.

``Nashpy`` is a Python library with all dependencies being part of the standard
scientific Python stack (numpy and scipy [@scipy]) thus it is portable. For
example Windows support is regularly tested through a Windows continuous
integration service (Appveyor).
```

I don't believe that there being another piece of software that has the same functionality precludes from publication in JOSS  but I might be mistaken.

Apologies again for the confusion and extra work that has resulted.
