### 2018-07-15 b69f05f
Add ability to consider supports of unequal size
