+++ tests/unit/test_support_enumeration.py
@@ -20,22 +20,107 @@ class TestSupportEnumeration(unittest.TestCase):
         A = np.array([[1, 0], [-2, 3]])
         B = np.array([[3, 2], [-1, 0]])
         self.assertEqual(list(potential_support_pairs(A, B)),
-                         [((0,), (0,)), ((0,), (1,)), ((1,), (0,)),
-                          ((1,), (1,)), ((0, 1), (0, 1))])
+                         [((0,), (0,)),
+                          ((0,), (1,)),
+                          ((0,), (0, 1)),
+                          ((1,), (0,)),
+                          ((1,), (1,)),
+                          ((1,), (0, 1)),
+                          ((0, 1), (0,)),
+                          ((0, 1), (1,)),
+                          ((0, 1), (0, 1))])
 
         A = np.array([[1, 0, 2], [-2, 3, 9]])
         B = np.array([[3, 2, 1], [-1, 0, 2]])
         self.assertEqual(list(potential_support_pairs(A, B)),
-                         [((0,), (0,)), ((0,), (1,)), ((0,), (2,)), ((1,), (0,)),
-                          ((1,), (1,)), ((1,), (2,)), ((0, 1), (0, 1)),
-                          ((0, 1), (0, 2)), ((0, 1), (1, 2))])
+                         [((0,), (0,)),
+                          ((0,), (1,)),
+                          ((0,), (2,)),
+                          ((0,), (0, 1)),
+                          ((0,), (0, 2)),
+                          ((0,), (1, 2)),
+                          ((0,), (0, 1, 2)),
+                          ((1,), (0,)),
+                          ((1,), (1,)),
+                          ((1,), (2,)),
+                          ((1,), (0, 1)),
+                          ((1,), (0, 2)),
+                          ((1,), (1, 2)),
+                          ((1,), (0, 1, 2)),
+                          ((0, 1), (0,)),
+                          ((0, 1), (1,)),
+                          ((0, 1), (2,)),
+                          ((0, 1), (0, 1)),
+                          ((0, 1), (0, 2)),
+                          ((0, 1), (1, 2)),
+                          ((0, 1), (0, 1, 2))])
 
         A = np.array([[1, 0], [-2, 3], [2, 1]])
         B = np.array([[3, 2], [-1, 0], [5, 2]])
         self.assertEqual(list(potential_support_pairs(A, B)),
-                         [((0,), (0,)), ((0,), (1,)), ((1,), (0,)),
-                          ((1,), (1,)), ((2,), (0,)), ((2,), (1,)),
-                          ((0, 1), (0, 1)), ((0, 2), (0, 1)), ((1, 2), (0, 1))])
+                         [((0,), (0,)),
+                          ((0,), (1,)),
+                          ((0,), (0, 1)),
+                          ((1,), (0,)),
+                          ((1,), (1,)),
+                          ((1,), (0, 1)),
+                          ((2,), (0,)),
+                          ((2,), (1,)),
+                          ((2,), (0, 1)),
+                          ((0, 1), (0,)),
+                          ((0, 1), (1,)),
+                          ((0, 1), (0, 1)),
+                          ((0, 2), (0,)),
+                          ((0, 2), (1,)),
+                          ((0, 2), (0, 1)),
+                          ((1, 2), (0,)),
+                          ((1, 2), (1,)),
+                          ((1, 2), (0, 1)),
+                          ((0, 1, 2), (0,)),
+                          ((0, 1, 2), (1,)),
+                          ((0, 1, 2), (0, 1))])
+
+    def test_potential_supports_with_non_degenerate_flag(self):
+        """Test for the enumeration of potential supports when constrained to
+        non degenerate games"""
+        A = np.array([[1, 0], [-2, 3]])
+        B = np.array([[3, 2], [-1, 0]])
[…]
