### 2018-07-16 d3a5183
Add warning when even number of equilibria are returned
This points to degeneracy.
