+++ tests/unit/test_game.py
@@ -92,17 +92,20 @@ Column player:
     def test_property_support_enumeration(self, A, B):
         """Property based test for the equilibria calculation"""
         g = nash.Game(A, B)
-        for equilibrium in g.support_enumeration():
-            for i, s in enumerate(equilibrium):
-                # Test that have a probability vector (subject to numerical
-                # error)
-                self.assertAlmostEqual(s.sum(), 1)
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            for equilibrium in g.support_enumeration():
+                for i, s in enumerate(equilibrium):
+                    # Test that have a probability vector (subject to numerical
+                    # error)
+                    self.assertAlmostEqual(s.sum(), 1)
 
-                # Test that it is of the correct size
-                self.assertEqual(s.size, [3, 4][i])
+                    # Test that it is of the correct size
+                    self.assertEqual(s.size, [3, 4][i])
 
-                # Test that it is non negative
-                self.assertTrue(all(s >= 0))
+                    # Test that it is non negative
+                    self.assertTrue(all(s >= 0))
+
 
     def test_support_enumeration_for_bi_matrix(self):
         """Test for the equilibria calculation support enumeration"""
@@ -149,6 +152,59 @@ Column player:
                 self.assertTrue(np.array_equal(s1, s2),
                                 msg="obtained: {} !=expected: {}".format(obtained,
                                                                          expected))
+    def test_support_enumeration_for_degenerate_bi_matrix_game(self):
+        """Test for the equilibria calculation support enumeration with a
+        degenerate game"""
+        A = np.array([[-1, 0], [-1, 1]])
+        B = np.array([[1, 0], [1, -1]])
+        g = nash.Game(A, B)
+        expected_equilibria = [(np.array([1, 0]), np.array([1, 0])),
+                               (np.array([0, 1]), np.array([1, 0]))]
+        with warnings.catch_warnings(record=True) as w:
+            obtained_equilibria = list(g.support_enumeration())
+            for obtained, expected in zip(obtained_equilibria,
+                                          expected_equilibria):
+                for s1, s2 in zip(obtained, expected):
+                    self.assertTrue(np.array_equal(s1, s2),
+                                    msg="obtained: {} !=expected: {}".format(obtained,
+                                                                             expected))
+            self.assertGreater(len(w), 0)
+            self.assertEqual(w[-1].category, RuntimeWarning)
+
+        A = np.array([[3, 3],[2, 5],[0, 6]])
+        B = np.array([[3, 3],[2, 6],[3, 1]])
+        g = nash.Game(A, B)
+        expected_equilibria = [(np.array([1, 0, 0]), np.array([1, 0])),
+                               (np.array([0, 1 / 3, 2 / 3]),
+                                np.array([1 / 3, 2 / 3]))]
+        with warnings.catch_warnings(record=True) as w:
+            obtained_equilibria = list(g.support_enumeration())
+            for obtained, expected in zip(obtained_equilibria,
+                                          expected_equilibria):
+                for s1, s2 in zip(obtained, expected):
+                    self.assertTrue(np.allclose(s1, s2),
+                                    msg="obtained: {} !=expected: {}".format(obtained,
+                                                                             expected))
+            self.assertGreater(len(w), 0)
+            self.assertEqual(w[-1].category, RuntimeWarning)
+
+        A = np.array([[0, 0],[0, 0]])
+        B = np.array([[0, 0],[0, 0]])
+        g = nash.Game(A, B)
+        expected_equilibria = [(np.array([1, 0]), np.array([1, 0])),
+                               (np.array([1, 0]), np.array([0, 1])),
+                               (np.array([0, 1]), np.array([1, 0])),
[…]
