### 2018-07-16 6827d46
Pass tol and degenerate arguments to support enum.
