### 2018-07-17 6371f8f
Write documentation for degeneracy.
