### 2018-07-17 77e5b0c
Add support for py3.7
