+++ .travis.yml
@@ -3,6 +3,7 @@ language: python
 python:
   - 3.5
   - 3.6
+  - 3.7
 
 cache:
   directories:
