### 2018-07-17 2a0a760
Support enumeration for degenerate games (#42)
* Add ability to consider supports of unequal size

* Add warning when even number of equilibria are returned

This points to degeneracy.

* Pass tol and degenerate arguments to support enum.

* Write documentation for degeneracy.
