+++ docs/reference/source/nashpy.algorithms.rst
@@ -1,0 +1,29 @@
+nash\.algorithms package
+========================
+
+Submodules
+----------
+
+nashpy\.algorithms\.support\_enumeration module
+-----------------------------------------------
+
+.. automodule:: nashpy.algorithms.support_enumeration
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+nashpy\.algorithms\.vertex\_enumeration module
+----------------------------------------------
+
+.. automodule:: nashpy.algorithms.vertex_enumeration
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+nashpy\.algorithms\.lemke\_howson module
+----------------------------------------
+
+.. automodule:: nashpy.algorithms.lemke_howson
+    :members:
+    :undoc-members:
+    :show-inheritance:
