+++ docs/reference/source/nash.algorithms.rst
@@ -1,29 +1,0 @@
-nash\.algorithms package
-========================
-
-Submodules
-----------
-
-nash\.algorithms\.support\_enumeration module
----------------------------------------------
-
-.. automodule:: nash.algorithms.support_enumeration
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-nash\.algorithms\.vertex\_enumeration module
---------------------------------------------
-
-.. automodule:: nash.algorithms.vertex_enumeration
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-nash\.algorithms\.lemke\_howson module
---------------------------------------
-
-.. automodule:: nash.algorithms.lemke_howson
-    :members:
-    :undoc-members:
-    :show-inheritance:

[…]
