### 2018-07-17 0612fc8
Add more alternative libraries. (#44)
