### 2018-07-23 #761 @arfon
> I would like to un-"withdraw" this paper, please let me know if it's simpler if I were to resubmit it perhaps?

OK, thanks for the heads up @drvinceknight. We should be able to work with this original submission.
