### 2018-07-23 #761 @arfon
@drvinceknight - looks like some of the LaTeX formatting is a little wonky in the compiled PDF ☝️. Could you try and fix that? If you make a change to your `paper.md` you can ask @whedon to recompile the paper with `@whedon generate pdf`
