### 2018-07-30 3f7b591
Fix compilation of paper.
