### 2018-07-30 60e5a0d
Remove unnecessary assertions.
