+++ tests/unit/test_game.py
@@ -210,20 +210,22 @@ Column player:
 
     def test_support_enumeration_for_deg_bi_matrix_game_with_non_deg(self):
 
-        A = np.array([[-1, 0, 1], [0, -1, 0], [1, -1, -1]])
+        A = np.array([[0, 0], [0, 0]])
         g = nash.Game(A)
         with warnings.catch_warnings(record=True) as w:
             obtained_equilibria = list(g.support_enumeration(non_degenerate=True))
+            self.assertEqual(len(obtained_equilibria), 4)
             self.assertGreater(len(w), 0)
             self.assertEqual(w[-1].category, RuntimeWarning)
 
 
     def test_support_enumeration_for_deg_bi_matrix_game_with_low_tol(self):
 
-        A = np.array([[-1, 0, 1], [0, -1, 0], [1, -1, -1]])
+        A = np.array([[0, 0], [0, 0]])
         g = nash.Game(A)
         with warnings.catch_warnings(record=True) as w:
             obtained_equilibria = list(g.support_enumeration(tol=0))
+            self.assertEqual(len(obtained_equilibria), 4)
             self.assertGreater(len(w), 0)
             self.assertEqual(w[-1].category, RuntimeWarning)
 
