### 2018-08-10 #761 @arfon
Hi @drvinceknight - things are pretty busy at JOSS right now (close to 100 submissions on the go!) so we're a little slow getting to things.

Can you look at [this list of potential reviewers](https://bit.ly/joss-reviewers) and suggest a few people that might be suitable to review your package?
