### 2018-08-10 #761 @drvinceknight
No problem at all, I completely understand: zero pressure/impatience on my side.

Looking through the list I'd suggest https://github.com/fil looks like a good fit, otherwise https://github.com/jordigh or https://github.com/nicoguaro perhaps.

Thank you again to all the editors for your efforts/work on this journal.
