### 2018-08-19 #904 @whedon
[REVIEW]: Nashpy: A Python library for the computation of Nash equilibria
**Submitting author:** @drvinceknight (<a href="http://orcid.org/0000-0002-4245-0638">Vincent Knight</a>)
**Repository:** <a href="https://github.com/drvinceknight/Nashpy" target ="_blank">https://github.com/drvinceknight/Nashpy</a>
**Version:** v0.0.18
