### 2018-08-19 #761 @arfon
> OK, I've started the review over in #904. Feel free to close this issue now!

@labarba - I found a weird bug with this paper. I've fixed it now.
