### 2018-08-19 drvinceknight/Nashpy#45 @Fil
review for JOSS
Looks good to me. A few details:

- [x] _Version: Does the release version given match the GitHub release (v0.0.14)?_
No: I see nashpy-0.0.17

- [x] _References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?_
Only one reference in paper.bib has a DOI.

- [x] Suggestions: I wonder if you could give more real-world usage examples and algorithms limitations. In particular, I tried with a few large random matrices 0 sum games and the game.support_enumeration() computation takes forever.

PS: testing at https://colab.research.google.com/drive/1UGrlXawmReLX4-y_vQqQeXhFEttfOWW9

(See https://github.com/openjournals/joss-reviews/issues/904)
