### 2018-08-20 bec046c
JOSS Review - Add DOIs (#47)
