+++ paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={48--49},
   year={1950},
-  publisher={USA}
+  publisher={USA},
+  doi={https://doi.org/10.1073/pnas.36.1.48},
 }
 
 @article{knight2017measuring,
@@ -17,7 +18,8 @@
   number={6},
   pages={630--642},
   year={2017},
-  publisher={Springer}
+  publisher={Springer},
+  doi={https://doi.org/10.1057},
 }
 
 @article{lemke1964equilibrium,
@@ -28,7 +30,8 @@
   number={2},
   pages={413--423},
   year={1964},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={https://doi.org/10.1137/0112033},
 }
 
 @article{mckelvey2006gambit,
