### 2018-08-20 2c0a74f
JOSS-Review add a discussion about algorithms.
Add some timings for an 11 by 11 game but also describe the fact that
the algorithms themselves are the limiting factor due to the complexity.
