### 2018-08-20 a31df7f
Make minor latex adjustments.
Use `$` instead of `\(` for pandoc compilation of LaTeX.
