+++ paper.md
@@ -97,16 +97,16 @@ software is automatically tested using a combination of doc (this paper is also
 tested), unit, integration and property based tests with 100% coverage.
 
 There are potential limitiations associated with Nashpy, these are due to the
-complexity of the algorithms themselves. For example, *support enumeration*
-enumerates all potential pairs of strategy. For \\(n\times n\\) sized square
-matrices it has complexity of the order of \\(O(2^n^2)\\). All implementations
-provided in Nashpy ensure these affects are reduced: numpy [@scipy] provides C
-based implementations for vectorized performance. Furthermore, all algorithms
-are generators which ensures that not all equilibria must be
-found before one is returned. For example, below an 11 by 11 game is
-considered and timings computed are shown for relative
-comparison.  Using the more efficient Lemke-Howson algorithm
-[@lemke1964equilibrium] an equilibria is founds approximately 3000 times faster.
+complexity of the algorithms themselves. For example, support enumeration
+enumerates all potential pairs of strategy. For $n\times n$ sized square
+matrices it has complexity of the order of $\mathcal{O}\left({2^n}^2\right)$.
+All implementations provided in Nashpy ensure these affects are reduced: numpy
+[@scipy] provides C based implementations for vectorized performance.
+Furthermore, all algorithms are generators which ensures that not all equilibria
+must be found before one is returned. For example, below an 11 by 11 game is
+considered and timings computed are shown for relative comparison.  Using the
+more efficient Lemke-Howson algorithm [@lemke1964equilibrium] an equilibria is
+founds approximately 3000 times faster.
 
 ```
 >>> from pprint import pprint
