### 2018-08-20 63b4e4e
JOSS-Review add a discussion about algorithms. (#48)
* JOSS-Review add a discussion about algorithms.

Add some timings for an 11 by 11 game but also describe the fact that
the algorithms themselves are the limiting factor due to the complexity.

* Make minor latex adjustments.

Use `$` instead of `\(` for pandoc compilation of LaTeX.
