### 2018-08-20 drvinceknight/Nashpy#45 @drvinceknight
Thank you very much for taking the time to do the review :+1:

> Version: Does the release version given match the GitHub release (v0.0.14)?
> No: I see nashpy-0.0.17

Great catch, thank you for checking this. I do not immediately see how to fix it. I have  updated Nashpy a few times since submitting the paper through joss's online system. I will comment on https://github.com/openjournals/joss-reviews/issues/904 to check how best to proceed. As that issue is opened by wheden (bot) I'm not entirely sure if it is as simple as me just editing the comment. For info:

- When I submitted the paper, the library was indeed at `v0.0.14` which is archived here: https://zenodo.org/record/1255027
- The latest release `v0.0.17` is archived here: https://zenodo.org/record/1313464

>
> References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
> Only one reference in paper.bib has a DOI.

I have addressed this here https://github.com/drvinceknight/Nashpy/pull/47

Note that two references do not have DOIs however I am referencing them as instructed in their documentation:

- Scipy: https://www.scipy.org/citing.html#scipy-the-library
- Gambit: see http://www.gambit-project.org/ under `Using and citing Gambit`

>
> Suggestions: I wonder if you could give more real-world usage examples and algorithms limitations. In particular, I tried with a few large random matrices 0 sum games and the game.support_enumeration() computation takes forever.

Very good idea. I hope to have addressed this satisfactorily in https://github.com/drvinceknight/Nashpy/pull/48. As I mention there the issue here is not as much the implementations (I believe they are implemented well) but the algorithms themselves: support enumeration is not an efficient algorithm (it is essentially brute force). In #48 I have:

- Added a large 11 by 11 game example
- Described the limitations of the algorithm
- Highlighted the fact that numpy ensures C bindings
- Highlighted the implementation choice of using generators which implies a user does not *need* to compute all equilibria to find *an* equilibria
- Included some timing measurements to highlight the difference in the algorithms themselves.

If you feel this could be done in a different/better I'd be happy to address it further.
