### 2018-08-20 drvinceknight/Nashpy#45 @Fil
I've updated the share link for the colab notebook: https://colab.research.google.com/drive/1UGrlXawmReLX4-y_vQqQeXhFEttfOWW9
