### 2018-08-20 drvinceknight/Nashpy#45 @Fil
I've addressed a few typos, but the addition is very clear. Thanks!
