### 2018-08-20 #904 @Fil
Apart from the version issue, it's all good for me. Thank you @drvinceknight for submitting to JOSS.
