### 2018-08-20 c084b2e
Fix typos helpfully pointed out by @Fil
