### 2018-08-20 drvinceknight/Nashpy#45 @drvinceknight
Thank you! I have addressed the typos (really appreciate you pointing them out, sorry I missed them!) in #49. I've left a quick answer to your point about the triviality of the game too, happy to discuss further if you deem it necessary :) :+1:
