### 2018-08-20 6022346
Fix typos helpfully pointed out by @Fil (#49)
