### 2018-08-20 #904 @labarba
I've updated the version number in the review issue.
