### 2018-09-10 cdbf97b
Fix two typos
