### 2018-09-10 f2aa390
Fix minor typo in README.
