### 2018-09-10 drvinceknight/Nashpy#51 @alex-konovalov
review 2 for JOSS
Hello @drvinceknight, here there are some minor comments regarding https://github.com/openjournals/joss-reviews/issues/904. Overall, it looks good!

1. The DOI at
https://github.com/drvinceknight/Nashpy/blob/60223465edc1b5fa51587ad4b5d7e8eac1f1ba1d/paper.bib#L22
can not be resolved

2. In all places, DOI is given as a URL, except here:
https://github.com/drvinceknight/Nashpy/blob/60223465edc1b5fa51587ad4b5d7e8eac1f1ba1d/paper.bib#L58
To be consistent, I suggest to stick to the same style.

3. I've corrected two typos at https://github.com/drvinceknight/Nashpy/pull/50

4. Just in case, do not forget in case you may have to mention any funding in acknowledgements.
