### 2018-09-10 #904 @alex-konovalov
@labarba review completed. Fixed two typos in https://github.com/drvinceknight/Nashpy/pull/50 and pointed out a couple of issues with DOIs at https://github.com/drvinceknight/Nashpy/issues/51. Otherwise, straightforward installation, proper testing and detailed documentation - @drvinceknight, thank you for submitting to JOSS!
