### 2018-09-10 drvinceknight/Nashpy#51 @drvinceknight
Thank you @alex-konovalov! I've opened #52 which addresses the doi's, will merge once CI has run (just to be sure).

There is no funding relevant here. (Thanks for the check though :+1:)
