+++ paper.bib
@@ -19,7 +19,7 @@
   pages={630--642},
   year={2017},
   publisher={Springer},
-  doi={https://doi.org/10.1057},
+  doi={https://doi.org/10.1057/s41274-016-0100-8},
 }
 
 @article{lemke1964equilibrium,
@@ -55,6 +55,6 @@
   title        = {drvinceknight/Nashpy: v0.0.13},
   month        = jan,
   year         = 2018,
-  doi          = {10.5281/zenodo.1163694},
+  doi          = {https://doi.org/10.5281/zenodo.1163694},
   url          = {https://doi.org/10.5281/zenodo.1163694}
 }
