### 2018-09-17 5d278c6
Run CI tests with pytest.
