### 2018-10-02 #904 @labarba
Small editorial fixes:

1. paragraph 3: "beats Rock (the first action." ... close parenthesis.
2. paragraph 5: "computing this equilibria" ... these equilibria? or this equilibrium? Same sentence: check commas; something isn't quite right.
3. page 2: "easily accessibly" ... accessible
4. p. 2, par. 2: capitalize "python" ... also, check verb: "includes ... to its core C functionality" (adds?) ... and commas.
5. par. 3: numpy > NumPy; scipy> SciPy; also, maybe use em-dashes to avoid the double parenthesis due to the citation.
6. For example Windows > add comma
7. par. 4: 2 player games > 2-player games (add hyphen)
8. par. 5: avoid syntactic expletive in "There are potential limitations ... , these are due to" > "Potential limitations of Nashpy are due to the complexity..."
9. "potential pairs of strategy" > a strategy?
10. "For n x n sized square matrices" > delete "sized"
11. "it has complexity of the order of O(2^n^2)" > it has O(.) complexity" (the big-O already means "order of")
12. numpy > NumPy
13. "are generators which ensures" ... comma before "which"
14. "For example, below an 11 by 11 game" > "For example, below, an 11-by-11 game" (comma and hyphens)
15. "and timings computed are shown" > "and timings are shown"
16. "Using the ... (ref) an equilibrium..." > comma after (ref)
17. "it is already currently" > delete "already"?
18. "final year course" > "final-year course" (hyphen)
19. "read and understand implementation of the algorithms" > the implementation
20. awkward: "with the linked DOI: (ref)" ... maybe just remove "with the linked DOI" and simply put the ref.
