### 2018-10-02 #904 @labarba
Also, DOI links in the reference list are broken. Please clean up the bib file so the links resolve.
