### 2018-10-05 #904 @drvinceknight
Thanks @labarba I'll make those changes over next couple of days.
