### 2018-10-08 67a27f3
Update paper.md following editorial comments
