+++ paper.bib
@@ -7,7 +7,7 @@
   pages={48--49},
   year={1950},
   publisher={USA},
-  doi={https://doi.org/10.1073/pnas.36.1.48},
+  doi={10.1073/pnas.36.1.48},
 }
 
 @article{knight2017measuring,
@@ -19,7 +19,7 @@
   pages={630--642},
   year={2017},
   publisher={Springer},
-  doi={https://doi.org/10.1057/s41274-016-0100-8},
+  doi={10.1057/s41274-016-0100-8},
 }
 
 @article{lemke1964equilibrium,
@@ -31,7 +31,7 @@
   pages={413--423},
   year={1964},
   publisher={SIAM},
-  doi={https://doi.org/10.1137/0112033},
+  doi={10.1137/0112033},
 }
 
 @article{mckelvey2006gambit,
@@ -55,6 +55,6 @@
   title        = {drvinceknight/Nashpy: v0.0.13},
   month        = jan,
   year         = 2018,
-  doi          = {https://doi.org/10.5281/zenodo.1163694},
+  doi          = {10.5281/zenodo.1163694},
   url          = {https://doi.org/10.5281/zenodo.1163694}
 }
