### 2018-10-08 #904 @drvinceknight
@labarba thank you very much for taking the time to point all those out. I've made all the changes (https://github.com/drvinceknight/Nashpy/commit/67a27f3300c804fa44ee393e7ff447c2bce126a0) and checked the proof. I believe it's all done now :)
