### 2018-10-08 #904 @labarba
I bring you more editorial fixes!

Page 1, par. 2: comma before “which”
par. 3 “For example A_21” …add comma. “Using Nashpy the equilibrium” …add comma.
Statement of Need: “an example of this includes” >> “an example is”
Page 2.
Ref to Jones, Oliphant et al. appears with years marked as 2001–2001– … can you fix this?
On the same line, remove spaces before and after the em-dashes (that’s the norm).
par. 5: “Nashpy” here is formatted differently than elsewhere—pick one: code or non-code font?
Jones et al. again appears with that strange year range. Can you fix?
