### 2018-10-09 af30119
Make further editorial fixes.
