### 2018-10-09 #904 @drvinceknight
Thanks again @labarba, I believe https://github.com/drvinceknight/Nashpy/commit/af30119a129995a66f2c7ef1ee9bf3d1ef1e4ce8 gets all these. If there's anything else let me know.

I appreciate you taking the time. :)
