### 2018-10-09 #904 @labarba
OK! Go ahead and make a deposit on Zenodo now (or your favorite archival repository) and do post the DOI here.
