### 2018-10-10 #904 @labarba
The author list on the Zenodo entry does not match the author list on the paper. You may need to manually edit the author list on Zenodo (which grabs that automatically from commit history, I think.)
