### 2018-10-10 #904 @labarba
👋 @arfon — this paper is accepted and ready to be published.

Many thanks to @Fil and @alex-konovalov for reviewing for JOSS. Without your efforts, this adventure in new-wave publishing would not be possible!
