### 2018-10-10 #904 @arfon
@drvinceknight - can you confirm that this maths looks OK? @whedon is giving me a warning when compiling the paper [10.21105.joss.00904.pdf](https://github.com/openjournals/joss-reviews/files/2465225/10.21105.joss.00904.pdf)

```
[WARNING] Could not convert TeX math '
  A=
  \begin{pmatrix}
  0  & -1 & 1  \\
  1  & 0  & -1 \\
  -1 & 1  & 0  \\
  \end{pmatrix}
  ', rendering as TeX```
