### 2018-10-10 #904 @drvinceknight
> @drvinceknight - can you confirm that this maths looks OK?

@arfon I confirm that that looks right. (I'd suggest that it looks like the compiler is just raising a warning before falling "up" to use a superset compiler with the amsmath library.)
