### 2018-10-10 #904 @drvinceknight
Thank you all once again for your time and effort. Having reviewed a few times it's been a pleasure to be on the other side. @arfon @labarba et al. well done for what you're doing here, this "new-wave publishing" is really enthusing.
